pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpxpy"
version = "0.0.1"
description = "Official Mathpix client for Python"
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name="Nico Jimenez", email="nicodjimenez@mathpix.com" },
  { name="Matt Chang", email="mattchang@mathpix.com" },
]
license = { file="LICENSE.txt" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "requests >= 2.20.0",
    "python-dotenv>=0.15.0",
]
keywords = ["mathpix", "mpx", "mpxpy", "py", "client", "ocr", "api", "sdk", "python","package", "development"]

# --- Project URLs (Links displayed on PyPI) ---
[project.urls]
Homepage = "https://github.com/Mathpix/mpxpy"
Repository = "https://github.com/Mathpix/mpxpy"
"Bug Tracker" = "https://github.com/Mathpix/mpxpy/issues"
[tool.setuptools.packages.find]
exclude = ["tests*"]